levelSetFront: working timeIndex writing
The levelSetFront::write(label) is used by levelSetFront::write(const
Foam::Time&) to write down the front based on the provided extension to
the constructor (as well as local instance and file name). The front is
written down as frontName-00000timeIndex.extension in the instance
directory at every time step.

Initial walking skeleton for the meshFrontCommunication is set.