ENH: Updated several parts of the solver lentFoam
Disabled the implicit update of the velocity and use the explicit reconstruction as in interFoam for now as there where no significant differences in the velocity field. Adapted the lentFoam structure to the latest interFoam version (e.g. make use of the centralized functionality provided by CorrectPhi.H). Removed some headers not needed.